Add a script to report locationText for the position of the system caret (#13060)
Closes #9147
Related to #7314

Summary of the issue:
Currently it is impossible to report locationText for the current caret position only for the position of the review cursor. This is problematic in several cases:

As described in Add alternative way to report name of the current Excel sheet which don't require moving review cursor into Excel window #9147 there are cases in which it is useful to query the name of the current sheet in MS Excel
In edit controls and / or web browsers sighted people can simply glance at the scroll bar to determine their position in the text
While for people working with review cursor following system caret current NVDA+Numpad delete is sufficient for these use cases when review cursor does not follow caret you need to move it to focus every time when you need to check your current position which wastes time, and more importantly makes your review position lost.

Description of how this pull request fixes the issue:
NVDA+numpad delete reports location of the caret or focused object if there is no caret by default.
Script for reporting review cursor position has been reassigned to NVDA+shift+Numpad delete - for most users there is no change as navigator follows focus and review follows caret by default.
The modified scripts are splitted into separate ones for querying location of the text / object as discussed in Split out global commands into distinct scripts #7314.