Warn about unknown command line arguments when starting NVDA (#13087)
Closes #12795

Summary of the issue:
When starting NVDA with unknown (not used by NVDA nor by add-ons)
command line parameters there is no warning.

Description of how this pull request fixes the issue:
Introduced a new extension point which accumulates decisions made by all handlers and allowed to register add-ons for command line arguments they want to use. Note that if the given add-on contains only app modules it has to bundle a small global plugin to make sure that it is loaded on NVDA's startup.
On NVDA startup user is warned if there are any arguments which are unknown to NVDA and not used by add-ons.