Fixed an issue preventing chat message to be reported in some versions of Skype for Business. (#12122)
Fixes #9295

Summary of the issue:
In Skype for Business (at least Office 2016, version 16.0.5095.1000), the incoming messages are not reported and an error occurs in the log instead.
This error show that the message has a value set to None whereas the code expects a string in order to parse it.
(cf. #9295 (comment) for more details)

Description of how this pull request fixes the issue:
If the message value is None, parse only the message name to split the content from author and timestamp parts. For this, we just look at ", , " separators.
If the message is not None, use as previously the message name and the message value to split author, content and timestamp of the message.