Make punctuation in the descriptions of scripts more consistent (#13137)
Punctuation in the descriptions of the new scripts added in PR #13060 was used inconsistently i.e. missing full stops and commas.

Description of how this pull request fixes the issue:
Fixed these inconsistencies by making sure that all full sentences ends with a full stop and added a missing comma.