Remove deprecated constants from languageHandler (#13144)
Removes code marked as deprecated in #12753

Summary of the issue:
As part of #12753 various constants in languageHandler were moved into an enum but kept at the module level for backwards compatibility. Since NVDA 2022.1 is a backwards compatibility breaking release it makes sense to remove them.

Description of how this pull request fixes the issue:
Deprecated constants are removed.