winVersion: remove deprecated Windows 10 releases to builds map (#13211)
Removes deprecated code from #12509

Summary of the issue:
As of #12544, Windows 10 releases to builds map was deprecated, with what's new document declaring that this map will be gone in 2022.1. This pull request accomplishes exactly that.

Description of how this pull request fixes the issue:
Removes deprecated Windows 10 releases to builds as Windows Registry is used to obtain Windows 10/11 release information.