Remove deprecated SYMLVL constants from characterProcessing (#13248)
Removes code deprecated as part of PR #11856

Summary of the issue:
PR #11856 converted module level constants for symbol levels in characterProcessing into an enum. The constants were marked for removal in 2022.1.

Description of how this pull request fixes the issue:
These constants are removed.

Testing strategy:
With git grep made sure that the removed symbol levels are not used in the source code.