Force DictionaryDialog to be subclassed to use it (#13268)
Link to issue number:
Follow-up of #12967

Summary of the issue:
Since the merge of #12800, trying to re-open an already opened dialog allows to focus it rather than issuing an error message box. The dialog is refocused according to its class, i.e. trying to open a dialog of the same class rather refocus the first already opened dialog of the same class.

But with #12800, it was not possible to open 2 dictionaries of different type (e.g. speech and default). This was fixed with #12967 by subclassing DictionaryDialog.
However, addons which directly use DictionaryDialog will still have the issue of opening 2 dictionaries at once i.e. #5383.

Description of how this pull request fixes the issue:
force DictionaryDialog to be subclassed by making the initializer and abstract method