Log WindowErrors for easeOfAccess and config (#13242)
Link to issue number:
Relates to #13155

Summary of the issue:
In order to investigate #13155, further logging is needed.

Description of how this pull request fixes the issue:
Adds logging for WindowErrors raised in easeOfAccess and config. This includes improvements to functions unrelated to #13155, which access the registry, by adding logging to prevent similar issues.