Support IDispatch returned for accSelection
Link to issue number:
Fixes #13276
(together with LibreOffice change
https://gerrit.libreoffice.org/c/core/+/129201 )

Summary of the issue:
In case a single child is selected,
'Accessible::get_accSelection' can either return
the child ID of a selected simple element or
an IDispatch for a selected accessible object, s.
https://docs.microsoft.com/en-us/windows/win32/api/oleacc/nf-oleacc-iaccessible-get_accselection
and
https://docs.microsoft.com/en-us/windows/win32/winauto/how-servers-implement-child-ids .

Previously, only positive child IDs were considered
in NVDA, while LibreOffice was returning negative child IDs
for Calc cells, which does not fit the MSAA spec.

NVDA did not support the (valid) case where an
IDispatch is returned.

Description of how this pull request fixes the issue:
This PR implements the missing handling for an
IDispatch returned as accSelection in NVDA and
LibreOffice returns such a one from LibreOffice
change https://gerrit.libreoffice.org/c/core/+/129201
on.