Show IAccessible2 relations in developer info (#13315)
Summary of the issue:
IAccessible2 has a concept of relations between objects. Among other things, this is used for aria-details, aria-controls, etc.

Description of how this pull request fixes the issue:
Show the relations for an object in the developer info for ease of debugging. Note that this only shows the relations, not the objects itself.

Testing strategy:
Used the example from https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/Attributes/aria-controls to prove that the relations shown in code are also visible in the developer info with this patch applied.