Fix system tests for source copies of NVDA (#13322)
Link to issue number:
None - Fixes regression introduced in PR #12075

Summary of the issue:
When executing system tests ensuring that NVDA starts and exits correctly we rely on the status of its process i.e. when the process is running NVDA is supposed to be running. While this works well for installed copies source copies are not started directly rather intermediate .bat file is used to start pythonw in a separate process. As a consequence when trying to retrieve status of an NVDA process during tests we get wrong results (the bat file is never running).

Description of how this pull request fixes the issue:
To avoid relying on the state of the process system tests use the NVDA's message window to determine if NVDA is running or not.