HID Braille driver: allow performing chorded gestures (space+dots) (#13326)
When using a Braille display in NVDA via HID braille, it was impossible to perform chorded gestures, that is space with other dots. E.g. space+dot4 to move down by line.

Description of how this pull request fixes the issue:
Fix an indenting error which caused only the last key in a chorded gesture to be listed in the BrailleGesture's id.