Reinitialize Freedom Scientific displays on I/O cancellation
Fixes #9830

After control of the session is returned to the user
from the post-resume lock screen, I/O directed at some Focus Blue
braille displays would raise an error about the operation being cancelled,
causing the I/O thread to crash but not triggering a restart
of the display driver, leading the display to not function correctly
until NVDA was restarted.

This commit adds a callback that allows the I/O exception
thrown in hwio to be supressed and uses it to terminate and
reinitialize the Braille display after resume.

One flaw with the current strategy is that the termination can
cause more I/O requests to be cancelled, leading to repeated
renitilization.