Prevent multiple settings dialogs opened (#13117)
Link to issue number:
Fixes #12818

Summary of the issue:
Two settings dialogs can be opened when using the steps in #12818.

Wx destroy events (and other events) are sent to parent dialogs for handling. 
When navigating to braille or speech settings, a ExpandoTextCtrl is initialized.
The ExpandoTextCtrl creates a destroy event as part of initialization, this caused the NVDASettings dialog to be incorrectly set to destroyed.

Description of how this pull request fixes the issue:
Filter the destroy event on NVDA Settings dialogs by checking if the dialog itself is being destroyed, as opposed to a child control.

Testing strategy:
system tests are used in the PR commit history to confirm the fix.
Manual testing of the steps in #12818 also confirms this fix.