Correctly reconnect to Focus Blue displays on resume (#13330)
Link to issue number:
Fixes #9830

Summary of the issue:
NVDA doesn't correctly regain control of some Focus Blue displays after the user has left the post-resume secure desktop. When in this state, Braille is only sent to the display when on a secure desktop. To regain normal functionality, the user must restart NVDA, possibly after unplugging the display and plugging it back in again.

From a technical perspective, the failure mode shows up as an error 995 (referring to cancelled I/O operations) in the background reader thread, which then leads to all the ack packets timing out.

Description of how this pull request fixes the issue:
This PR adds an onReadError callback to hwIo, which allows hwIo users to handle I/O errors that can't be caught with a try/catch (since the exception happens on a different thread).

It then uses this callback to detect the interrupted I/O characteristic of the post-resume failure state. When the issue occurs, the driver is terminated and reinitialized.

Testing strategy:
I suspended and resumed my computer while running a development build that included this fix and observed that, unlike with the latest release, the Braille display reflected changes to what I'd typed in a terminal after resume, instead of consistently showing "Secure Desktop".

I also checked the logs to ensure that the driver was only initialized once and that there weren't other exceptions being thrown as a result of this change.

* Reinitialize Freedom Scientific displays on I/O cancellation

Fixes #9830

After control of the session is returned to the user
from the post-resume lock screen, I/O directed at some Focus Blue
braille displays would raise an error about the operation being cancelled,
causing the I/O thread to crash but not triggering a restart
of the display driver, leading the display to not function correctly
until NVDA was restarted.

This commit adds a callback that allows the I/O exception
thrown in hwio to be supressed and uses it to terminate and
reinitialize the Braille display after resume.

One flaw with the current strategy is that the termination can
cause more I/O requests to be cancelled, leading to repeated
renitilization.

* Lint

* Clean up the logic, go for better function names

* update changes

Co-authored-by: buddsean <sean@nvaccess.org>