No longer try to detect spelling errors outside of editable content. (#13247)
Fixes #13051

Summary of the issue:
During NVDA 2021.3 development cycle we started logging exceptions which occur when we cannot detect if the previous word contains spelling errors. The detection code was placed in the base NVDAObject and therefore was triggered when it should not i.e. when checking checkboxes in NVDA's preferences. This causes a unnecessary noise in the log.

Description of how this pull request fixes the issue:
The detection of spelling errors has been moved from the base NVDAObjects to the NVDAObjects.behaviors.EditableText and therefore is never triggered outside of the editable text.

Testing strategy:
Ensured that spelling errors and typed characters are properly reported in MS Word with the object model, Libre Office Writer and in Firefox (both in the standard edit fields and in the compound documents e.g. Etherpad).