Disallow opening the browse for addons dialog on secure screens to avoid a security exploit. (#13059)
Summary of the issue:
It was possible for a user to bind a gesture to open the addons manager, copy their config to secure screens, press it, press the install button, and get a browse dialog, allowing them to run CMD as systemroot, and do all sorts of things.

Description of how this pull request fixes the issue:
Only show the addon manager if we're not on a secure screen.