Disable context help in secure screen to avoid a security exploit. (#13353)
If you open NVDA parameter from secure screen and press F1, context help opens in Internet explorer.
I am not an expert in security but I think that it's not secure to have access to the browser from secure screen.
More specifically, via the browser open dialog, I am able to open any file in notepad and modify it, e.g. allowing me to activate NVDA console.

Description of how this pull request fixes the issue:
Since context help is displayed in a browser, disable it on secure screen.