Disable GUI inspection tool in secure mode
GitHub Advisory GHSA-mvc8-5rv9-w3hx

Summary:
The wx GUI inspection tool includes a python console.
If the user binds a gesture to the startWxInspectionTool script and
their config is copied to be used on logon screen, this tool could then be
opened from the logon screen.
This would allow a user to open the python console from the logon screen with
system privileges.

Description of change:
Disables opening the wx GUI inspection tool when NVDA is
running in secure mode.