No longer use fileinput when generating COMInterfaces as it breaks parallel builds (#13367)
Fixes #13290
Related to this thread on nvda-devel by @MarcoZehe

Summary of the issue:
PR #13226 started building nVDA in parallel. While this offers significant speed improvements the way in which we were generating COMInterfaces was not thread safe. The issue was caused by the fileinput module which redirects stdout to a given file and this cannot be done in multiple threads at the same time. Also, as discovered by @MarcoZehe in the thread linked above when stdout is redirected to a file some other build step which is executing in a separate thread can write to it and the text which should be shown on screen ends up in the COMInterface file.

Description of how this pull request fixes the issue:
Were not using fileinput anymore - the IDE friendly content is generated in memory and then written to the file afterwards.