Upgrade to py2exe 0.11.1.0 (#13510)
Some time after we branched for the 2021.3 beta, pr #13066 was merged to master, which upgraded Py2exe to 0.11.0.1. However, this version of Py2exe fails to build the NVDA distribution when using optimized Python (E.g. when building a release).
This was not noticed until trying to build the NVDA 2022.1beta1 tag.

Description of how this pull request fixes the issue:
This pr upgrades py2exe to 0.11.1.0.