Fix table navigation commands through merged cells by caching last column/row index (#13345)
Follow up on the work done by @mltony in #11923

Fixes #7278, #11919

Summary of the issue:
When issuing table navigation commands repeatedly, e.g. pressing Control+Alt+Down several times in a row, the original column index is currently not preserved. Thus, if there happens to be a merged cell on the way, the column index will be reset to the minimal column index of merged cell.

Description of how this pull request fixes the issue:
This PR caches original column index for vertical table navigation and original row index for horizontal table navigation.